+++ script/character/class_perk.ts
@@ -35,7 +35,7 @@ class ClassPerk {
     if (this.unlock.stats) {
       const stats = this.unlock.stats;
       const pStats = player.getStats();
-      stats.forEach((stat: string) => {
+      Object.keys(stats).forEach((stat: string) => {
         if (pStats[stat] < stats[stat]) return false;
       });
     }
@@ -76,9 +76,22 @@ class ClassPerk {
     }
 
     // Perk description
-    tooltip += `<c>silver<c>${game.getLocalizedString(this.id) + "_desc"}<c>white<c>\n`;
+    tooltip += `<c>silver<c>"${game.getLocalizedString(this.id + "_desc")}"<c>white<c>\n`;
+
+    if (player.hasClassPerk(this.id)) {
+      tooltip += `<c>lime<c>${game.getLocalizedString("class_perk_owned")}<c>white<c>\n`;
+    }
 
     // Perk unlock
+    if (this.unlock.stats) {
+      tooltip += `<c>silver<c>Requires:\n`;
+      Object.entries(this.unlock.stats).forEach(([key, value]: any) => {
+        const statCol = player.getStats()[key] >= value ? "lime" : "red";
+        tooltip += ` <i>${icons[key]}<i>${game.getLocalizedString(key)}: <c>${statCol}<c>${value}<c>white<c>\n`;
+      });
+    }
+    const priceCol = player.gold > this.unlock.gold ? "gold" : "red";
+    tooltip += `Price: <c>${priceCol}<c>${compactNumber(this.unlock.gold)}<c>white<c>\n`;
 
     if (this.commands) {
       Object.entries(this.commands).forEach(([key, value]: [string, any]) => {
@@ -97,7 +110,6 @@ class ClassPerk {
         tooltip += " " + effectSyntax(key, value);
       });
     }
-    console.log(tooltip);
     return tooltip;
   }
 }
@@ -117,9 +129,135 @@ interface ClassPerkGroup {
 }
 
 const classPerks: ClassPerks = {
-  warrior: [],
-  rogue: [],
-  mage: [],
+  warrior: [
+    {
+      level: 1,
+      perks: [
+        new ClassPerk({
+          id: "warrior_vitality",
+          type: "classPerk",
+          class: "warrior",
+          unlock: {
+            stats: {
+              str: 15,
+            },
+            gold: 250,
+          },
+          modifiers: {
+            hpMaxFromStrV: 0.5,
+          },
+        }),
+        new ClassPerk({
+          id: "warrior_power",
+          type: "classPerk",
+          class: "warrior",
+          unlock: {
+            gold: 100,
+          },
+          modifiers: {
+            meleeDamageP: 5,
+          },
+        }),
+        new ClassPerk({
+          id: "warrior_strength",
+          type: "classPerk",
+          class: "warrior",
+          unlock: {
+            gold: 100,
+          },
+          modifiers: {
+            strP: 5,
+          },
+        }),
+      ],
+    },
+  ],
+  rogue: [
+    {
+      level: 1,
+      perks: [
+        new ClassPerk({
+          id: "rogue_make_haste",
+          type: "classPerk",
+          class: "rogue",
+          unlock: {
+            stats: {
+              agi: 15,
+            },
+            gold: 250,
+          },
+          commands: {
+            add_ability: { ...abilities.haste_enchant },
+          },
+        }),
+        new ClassPerk({
+          id: "rogue_nimble",
+          type: "classPerk",
+          class: "rogue",
+          unlock: {
+            gold: 100,
+          },
+          modifiers: {
+            speedP: 5,
+          },
+        }),
+        new ClassPerk({
+          id: "rogue_agile",
+          type: "classPerk",
+          class: "rogue",
+          unlock: {
+            gold: 100,
+          },
+          modifiers: {
+            agiP: 5,
+          },
+        }),
+      ],
+    },
+  ],
+  mage: [
+    {
+      level: 1,
+      perks: [
+        new ClassPerk({
+          id: "mage_mana_regen",
+          type: "classPerk",
+          class: "mage",
+          unlock: {
+            stats: {
+              int: 15,
+            },
+            gold: 250,
+          },
+          modifiers: {
+            mpRegenP: 10,
+          },
+        }),
+        new ClassPerk({
+          id: "mage_spell_mastery",
+          type: "classPerk",
+          class: "mage",
+          unlock: {
+            gold: 100,
+          },
+          modifiers: {
+            spellPowerP: 5,
+          },
+        }),
+        new ClassPerk({
+          id: "mage_books",
+          type: "classPerk",
+          class: "mage",
+          unlock: {
+            gold: 100,
+          },
+          modifiers: {
+            intP: 5,
+          },
+        }),
+      ],
+    },
+  ],
   paladin: [
     {
       level: 1,
@@ -129,10 +267,36 @@ const classPerks: ClassPerks = {
           type: "classPerk",
           class: "paladin",
           unlock: {
-            gold: 100,
+            stats: {
+              spi: 15,
+            },
+            gold: 250,
           },
           commands: {
             add_ability: { ...abilities.smite },
+          },
+        }),
+        new ClassPerk({
+          id: "paladin_vigour",
+          type: "classPerk",
+          class: "paladin",
+          unlock: {
+            gold: 100,
+          },
+          modifiers: {
+            hpMaxP: 3,
+            hpMaxV: 10,
+          },
+        }),
+        new ClassPerk({
+          id: "paladin_strength",
+          type: "classPerk",
+          class: "paladin",
+          unlock: {
+            gold: 100,
+          },
+          modifiers: {
+            strP: 5,
           },
         }),
       ],