+++ script/character/player.ts
@@ -363,6 +363,10 @@ class Player extends Character {
   hasKeyItem(item: string): boolean {
     return this.key_items?.includes(item);
   }
+
+  hasClassPerk(perk: string): boolean {
+    return this.class.perks?.find((p) => p.id === perk) !== undefined;
+  }
 }
 
 const defaultPlayer = {
@@ -413,7 +417,7 @@ const defaultPlayer = {
   level: 1,
   xp: 0,
   starting_aspect: "strength",
-  class: classManager.get("paladin"),
+  class: classManager.get("warrior"),
 };
 
 let player = new Player({