+++ script/compiled/ability/effects.js
@@ -197,5 +197,16 @@ const effects = {
             damageP: 10,
         },
     },
+    haste_1: {
+        id: "haste_1",
+        icon: "gfx/icons/running-ninja.png",
+        duration: 7,
+        type: "buff",
+        buffDebuff: "buff",
+        modifiers: {
+            speedP: 10,
+            dodgeP: 10,
+        },
+    },
 };
 //# sourceMappingURL=effects.js.map