+++ script/compiled/lobby/class_view.js
@@ -26,7 +26,10 @@ function createClassView() {
             perkElem.classList.add("class-perk");
             perkElem.innerText = game.getLocalizedString(perk.id);
             tooltip(perkElem, perk.tooltip());
-            if (!perk.available()) {
+            if (player.hasClassPerk(perk.id)) {
+                perkElem.classList.add("owned");
+            }
+            else if (!perk.available()) {
                 perkElem.classList.add("locked");
             }
             perkElem.onclick = () => {