+++ script/language/general.ts
@@ -130,6 +130,9 @@ const english = {
   level: "Level",
   confirm_perk_unlock: `Are you sure you want to unlock perk "{id}" ?`,
 
+  // Special
+  divine_desc: "Extra damage vs unholy enemies",
+
   // Menu
   new_game: "New Game",
   load_game: "Load Game",
@@ -428,6 +431,36 @@ const english = {
   will_7: "Will VII",
   will_8: "Will VIII",
   will_enlightened: "Will Enlightened",
+
+  // Class specific perks
+  // WARRIOR
+  warrior_vitality: "Warrior's Vitality",
+  warrior_vitality_desc: "Close combat invigorates you.",
+  warrior_power: "Combat Prowess",
+  warrior_power_desc: "Your skill in melee improves.",
+  warrior_strength: "Strength Training",
+  warrior_strength_desc: "Those muscles won't build themselves.",
+  // ROGUE
+  rogue_make_haste: "Make Haste",
+  rogue_make_haste_desc: "Keep moving or perish.",
+  rogue_nimble: "Nimbler",
+  rogue_nimble_desc: "Your foes are struggling to keep track of you.",
+  rogue_agile: "Agility Training",
+  rogue_agile_desc: "You're getting better at dodging.",
+  // MAGE
+  mage_mana_regen: "Mana Regeneration",
+  mage_mana_regen_desc: "Your mana regenerates faster.",
+  mage_spell_mastery: "Spell Mastery",
+  mage_spell_mastery_desc: "Your spells are more powerful.",
+  mage_books: "Hitting the Books",
+  mage_books_desc: "Book good.",
+  // PALADIN
+  paladin_smite: "Smite the Wicked",
+  paladin_smite_desc: "Unholy beings will be destroyed before the divine.",
+  paladin_vigour: "Vigour",
+  paladin_vigour_desc: "Your flesh is blessed.",
+  paladin_strength: "Strength Training",
+  paladin_strength_desc: "Those muscles won't build themselves.",
 
   // Log
   reached_level: "Reached level {0}!",